+++ src/main/java/kitchenpos/menu/dto/IdOfOrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 public class IdOfOrderTableRequest {
     private Long id;
