+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -16,8 +16,8 @@ import java.util.List;
 import java.util.stream.Stream;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.menugroup.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuProductRequest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.dto.MenuProductRequest;
+import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.order.dto.OrderResponse;
