+++ src/test/java/kitchenpos/acceptance/TableAcceptanceTest.java
@@ -16,8 +16,8 @@ import java.util.Arrays;
 import java.util.stream.Stream;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.menugroup.dto.MenuGroupResponse;
-import kitchenpos.dto.MenuProductRequest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.dto.MenuProductRequest;
+import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.dto.OrderTableRequest;
 import kitchenpos.dto.OrderTableResponse;
