+++ src/test/java/kitchenpos/domain/MenuTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.math.BigDecimal;
+import kitchenpos.menu.domain.Menu;
 import org.junit.jupiter.api.Test;
 
 class MenuTest {
